ai_service/requirements.txt
fastapi>=0.103.0
uvicorn>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.4.0
requests>=2.31.0
python-multipart>=0.0.6
aiohttp>=3.8.5
openai>=1.0.0